+++ .gitignore
@@ -4,3 +4,4 @@ package-lock.json
 /*.tgz
 DEBUG-*
 /.yarn
+/dist
