+++ eslint.config.mjs
@@ -1,0 +1,5 @@
+import { generateEslintConfig } from '@companion-module/tools/eslint/config.mjs'
+
+export default generateEslintConfig({
+	enableTypescript: true,
+})
