+++ package.json
@@ -1,10 +1,16 @@
 {
 	"name": "lgtv-display",
 	"version": "2.0.0",
-	"main": "index.js",
+	"type": "module",
+	"main": "dist/main.js",
 	"scripts": {
 		"format": "prettier -w .",
-		"package": "companion-module-build"
+		"build": "rimraf dist && yarn build:main",
+		"build:main": "tsc -p tsconfig.build.json",
+		"dev": "tsc -p tsconfig.build.json --watch",
+		"lint:raw": "eslint",
+		"lint": "yarn lint:raw .",
+		"package": "yarn build && companion-module-build"
 	},
 	"license": "MIT",
 	"repository": {
@@ -12,17 +18,21 @@
 		"url": "git+https://github.com/bitfocus/companion-module-lgtv-display.git"
 	},
 	"engines": {
-		"node": "22",
-		"yarn": "4"
+		"node": "^22.20",
+		"yarn": "^4"
 	},
 	"dependencies": {
-		"@companion-module/base": "^1.14.1",
+		"@companion-module/base": "~1.14.1",
 		"lgtv-ip-control": "^4.4.0"
 	},
 	"devDependencies": {
 		"@companion-module/tools": "^2.7.2",
+		"@types/node": "^22.19.17",
 		"eslint": "^9.39.4",
-		"prettier": "^3.8.4"
+		"prettier": "^3.8.4",
+		"rimraf": "^6.1.3",
+		"typescript": "~5.9.3",
+		"typescript-eslint": "^8.59.0"
 	},
 	"prettier": "@companion-module/tools/.prettierrc.json",
 	"packageManager": "yarn@4.17.0"
