--- src/feedbacks.js
@@ -1,344 +1,0 @@
-const { combineRgb } = require('@companion-module/base')
-const { Apps, PowerStates } = require('lgtv-ip-control')
-const { appNames, inputNames, toChoices } = require('./lookups')
-
-const POLL_MIN = 250
-const POLL_MAX = 60000
-const POLL_DEFAULT = 2000
-
-module.exports = {
-	initFeedbacks: function () {
-		let self = this
-		let feedbacks = {}
-
-		feedbacks.powerState = {
-			type: 'boolean',
-			name: 'Power state',
-			description: 'True when the TV power state matches the selected state',
-			defaultStyle: {
-				color: combineRgb(255, 255, 255),
-				bgcolor: combineRgb(0, 153, 0),
-			},
-			options: [
-				{
-					type: 'dropdown',
-					id: 'state',
-					label: 'State',
-					default: PowerStates.on,
-					choices: [
-						{ id: PowerStates.on, label: 'On' },
-						{ id: PowerStates.off, label: 'Off' },
-						{ id: PowerStates.unknown, label: 'Unknown' },
-					],
-				},
-			],
-			callback: function (feedback) {
-				return self.feedbackState?.powerState === feedback.options.state
-			},
-		}
-
-		feedbacks.muteState = {
-			type: 'boolean',
-			name: 'Mute state',
-			description: 'True when mute state matches the selected state',
-			defaultStyle: {
-				color: combineRgb(255, 255, 255),
-				bgcolor: combineRgb(0, 153, 0),
-			},
-			options: [
-				{
-					type: 'dropdown',
-					id: 'muted',
-					label: 'Muted',
-					default: 'true',
-					choices: [
-						{ id: 'true', label: 'Muted' },
-						{ id: 'false', label: 'Unmuted' },
-					],
-				},
-			],
-			callback: function (feedback) {
-				return String(self.feedbackState?.isMuted ?? false) === feedback.options.muted
-			},
-		}
-
-		feedbacks.ipControl = {
-			type: 'boolean',
-			name: 'IP Control enabled',
-			description: 'True when GET_IPCONTROL_STATE returns ON',
-			defaultStyle: {
-				color: combineRgb(255, 255, 255),
-				bgcolor: combineRgb(0, 153, 0),
-			},
-			options: [],
-			callback: function () {
-				return self.feedbackState?.ipControlEnabled === true
-			},
-		}
-
-		feedbacks.currentVolume = {
-			type: 'boolean',
-			name: 'Volume compare',
-			description: 'Compares current volume against a value',
-			defaultStyle: {
-				color: combineRgb(255, 255, 255),
-				bgcolor: combineRgb(0, 153, 0),
-			},
-			options: [
-				{
-					type: 'dropdown',
-					id: 'op',
-					label: 'Operator',
-					default: 'eq',
-					choices: [
-						{ id: 'eq', label: '=' },
-						{ id: 'ne', label: '!=' },
-						{ id: 'gt', label: '>' },
-						{ id: 'gte', label: '>=' },
-						{ id: 'lt', label: '<' },
-						{ id: 'lte', label: '<=' },
-					],
-				},
-				{
-					type: 'number',
-					id: 'volume',
-					label: 'Volume (0-100)',
-					default: 50,
-					min: 0,
-					max: 100,
-					required: true,
-				},
-			],
-			callback: function (feedback) {
-				const currentVolume = self.feedbackState?.currentVolume
-				if (typeof currentVolume !== 'number') {
-					return false
-				}
-
-				const targetVolume = Number(feedback.options.volume)
-				switch (feedback.options.op) {
-					case 'eq':
-						return currentVolume === targetVolume
-					case 'ne':
-						return currentVolume !== targetVolume
-					case 'gt':
-						return currentVolume > targetVolume
-					case 'gte':
-						return currentVolume >= targetVolume
-					case 'lt':
-						return currentVolume < targetVolume
-					case 'lte':
-						return currentVolume <= targetVolume
-					default:
-						return false
-				}
-			},
-		}
-
-		feedbacks.currentApp = {
-			type: 'boolean',
-			name: 'Current app',
-			description: 'True when the current app matches the selected app',
-			defaultStyle: {
-				color: combineRgb(255, 255, 255),
-				bgcolor: combineRgb(0, 153, 0),
-			},
-			options: [
-				{
-					type: 'dropdown',
-					id: 'app',
-					label: 'App',
-					default: Apps.netflix,
-					choices: toChoices(appNames, 'Custom App ID'),
-				},
-				{
-					type: 'textinput',
-					id: 'customAppId',
-					label: 'Custom App ID',
-					default: '',
-					isVisibleExpression: `$(options:app) == '__custom__'`,
-				},
-			],
-			callback: function (feedback) {
-				const currentApp = String(self.feedbackState?.currentApp ?? '').trim()
-				const selectedApp =
-					feedback.options.app === '__custom__'
-						? String(feedback.options.customAppId ?? '').trim()
-						: String(feedback.options.app ?? '').trim()
-
-				if (!selectedApp || !currentApp) {
-					return false
-				}
-
-				return currentApp === selectedApp
-			},
-		}
-
-		feedbacks.currentInput = {
-			type: 'boolean',
-			name: 'Current input',
-			description: 'True when the current physical input matches the selected input',
-			defaultStyle: {
-				color: combineRgb(255, 255, 255),
-				bgcolor: combineRgb(0, 153, 0),
-			},
-			options: [
-				{
-					type: 'dropdown',
-					id: 'input',
-					label: 'Input',
-					default: 'com.webos.app.hdmi1',
-					choices: toChoices(inputNames, 'Custom Input ID'),
-				},
-				{
-					type: 'textinput',
-					id: 'customInputId',
-					label: 'Custom Input ID',
-					default: '',
-					isVisibleExpression: `$(options:input) == '__custom__'`,
-				},
-			],
-			callback: function (feedback) {
-				const currentApp = String(self.feedbackState?.currentApp ?? '').trim()
-				const selectedInput =
-					feedback.options.input === '__custom__'
-						? String(feedback.options.customInputId ?? '').trim()
-						: String(feedback.options.input ?? '').trim()
-
-				if (!selectedInput || !currentApp) {
-					return false
-				}
-
-				return currentApp === selectedInput
-			},
-		}
-
-		feedbacks.signalPresent = {
-			type: 'boolean',
-			name: 'Signal present',
-			description: 'True when the current input is receiving a signal (physical inputs only)',
-			defaultStyle: {
-				color: combineRgb(255, 255, 255),
-				bgcolor: combineRgb(0, 153, 0),
-			},
-			options: [],
-			callback: function () {
-				return self.feedbackState?.signal === true
-			},
-		}
-
-		self.setFeedbackDefinitions(feedbacks)
-		self.checkFeedbacks()
-	},
-
-	startFeedbackPolling: function () {
-		let self = this
-		self.stopFeedbackPolling()
-
-		const requested = Math.floor(Number(self.config?.poll_interval))
-
-		// A value of 0 disables periodic polling. Still run a single update so
-		// feedbacks/variables reflect the current state right after connecting.
-		if (requested === 0) {
-			self.updateFeedbackState()
-			return
-		}
-
-		const interval =
-			Number.isFinite(requested) && requested >= POLL_MIN && requested <= POLL_MAX ? requested : POLL_DEFAULT
-
-		self.feedbackPollTimer = setInterval(() => {
-			self.updateFeedbackState()
-		}, interval)
-		self.updateFeedbackState()
-	},
-
-	stopFeedbackPolling: function () {
-		let self = this
-		if (self.feedbackPollTimer) {
-			clearInterval(self.feedbackPollTimer)
-			self.feedbackPollTimer = undefined
-		}
-		self.feedbackPollInFlight = false
-	},
-
-	updateFeedbackState: async function () {
-		let self = this
-		// A single TCP socket serves all commands with no internal queue, so two
-		// overlapping polls would interleave their reads. Skip if one is running.
-		if (self.feedbackPollInFlight) {
-			return
-		}
-		self.feedbackPollInFlight = true
-
-		try {
-			if (!self.lgtv || !self.lgtv.connected) {
-				self.feedbackState = {
-					...self.feedbackState,
-					powerState: PowerStates.unknown,
-					currentApp: '',
-					currentVolume: null,
-					isMuted: false,
-					ipControlEnabled: false,
-					signal: undefined,
-					hdcpVersion: '',
-					hdcpStatus: '',
-				}
-				self.checkFeedbacks()
-				self.checkVariables()
-				return
-			}
-
-			let nextState = { ...self.feedbackState }
-
-			// getCurrentAppDetails runs the same CURRENT_APP command as getCurrentApp but
-			// also returns signal/HDCP info, so we get those for no extra round trip.
-			try {
-				const details = await self.lgtv.getCurrentAppDetails()
-				if (details === null) {
-					nextState.currentApp = ''
-					nextState.powerState = PowerStates.off
-					nextState.signal = undefined
-					nextState.hdcpVersion = ''
-					nextState.hdcpStatus = ''
-				} else {
-					nextState.currentApp = details.app ?? ''
-					nextState.powerState = PowerStates.on
-					nextState.signal = details.signal
-					nextState.hdcpVersion = details.hdcpVersion ?? ''
-					nextState.hdcpStatus = details.hdcpStatus ?? ''
-				}
-			} catch (error) {
-				nextState.currentApp = ''
-				nextState.powerState = PowerStates.unknown
-				nextState.signal = undefined
-				nextState.hdcpVersion = ''
-				nextState.hdcpStatus = ''
-			}
-
-			try {
-				nextState.currentVolume = await self.lgtv.getCurrentVolume()
-			} catch (error) {
-				nextState.currentVolume = null
-			}
-
-			try {
-				nextState.isMuted = await self.lgtv.getMuteState()
-			} catch (error) {
-				nextState.isMuted = false
-			}
-
-			try {
-				nextState.ipControlEnabled = await self.lgtv.getIpControlState()
-			} catch (error) {
-				nextState.ipControlEnabled = false
-			}
-
-			self.feedbackState = nextState
-			self.checkFeedbacks()
-			self.checkVariables()
-		} finally {
-			self.feedbackPollInFlight = false
-		}
-	},
-}
