+++ src/upgrades.ts
@@ -1,16 +1,29 @@
-const { Keys } = require('lgtv-ip-control')
+import type {
+	CompanionUpgradeContext,
+	CompanionStaticUpgradeProps,
+	CompanionStaticUpgradeResult,
+	CompanionStaticUpgradeScript,
+} from '@companion-module/base'
+import { Keys } from 'lgtv-ip-control'
+import type { ModuleConfig } from './config.js'
 
-module.exports = [
-	function (context, props) {
-		// This is a placeholder than now cannot be used/removed
+export const UpgradeScripts: CompanionStaticUpgradeScript<ModuleConfig>[] = [
+	function (
+		_context: CompanionUpgradeContext<ModuleConfig>,
+		_props: CompanionStaticUpgradeProps<ModuleConfig>,
+	): CompanionStaticUpgradeResult<ModuleConfig> {
+		// This is a placeholder that now cannot be used/removed
 		return {
 			updatedConfig: null,
 			updatedActions: [],
 			updatedFeedbacks: [],
 		}
 	},
-	function (context, props) {
-		const result = {
+	function (
+		_context: CompanionUpgradeContext<ModuleConfig>,
+		props: CompanionStaticUpgradeProps<ModuleConfig>,
+	): CompanionStaticUpgradeResult<ModuleConfig> {
+		const result: CompanionStaticUpgradeResult<ModuleConfig> = {
 			updatedConfig: null,
 			updatedActions: [],
 			updatedFeedbacks: [],
@@ -18,9 +31,10 @@ module.exports = [
 
 		// The feedback polling feature added a "poll_interval" config field. Existing
 		// instances have no value for it, so seed the same default the config field uses.
-		if (props.config && props.config.poll_interval === undefined) {
-			props.config.poll_interval = 2000
-			result.updatedConfig = props.config
+		const config = props.config as (ModuleConfig & Record<string, unknown>) | null
+		if (config && config.poll_interval === undefined) {
+			config.poll_interval = 2000
+			result.updatedConfig = config
 		}
 
 		// Legacy "Send Key" actions stored the Keys property name (e.g. "volumeUp")
@@ -32,8 +46,8 @@ module.exports = [
 			}
 
 			const requestedKey = action.options.key
-			if (requestedKey !== undefined && Keys[requestedKey] !== undefined) {
-				action.options.key = Keys[requestedKey]
+			if (typeof requestedKey === 'string' && requestedKey in Keys) {
+				action.options.key = Keys[requestedKey as keyof typeof Keys]
 				result.updatedActions.push(action)
 			}
 		}
