+++ tsconfig.build.json
@@ -1,0 +1,14 @@
+{
+	"extends": "@companion-module/tools/tsconfig/node22/recommended",
+	"include": ["src/**/*.ts"],
+	"exclude": ["node_modules/**", "src/**/*spec.ts", "src/**/__tests__/*", "src/**/__mocks__/*"],
+	"compilerOptions": {
+		"outDir": "./dist",
+		"baseUrl": "./",
+		"paths": {
+			"*": ["./node_modules/*"]
+		},
+		"module": "Node16",
+		"moduleResolution": "Node16"
+	}
+}
