+++ tsconfig.json
@@ -1,0 +1,8 @@
+{
+	"extends": "./tsconfig.build.json",
+	"include": ["src/**/*.ts"],
+	"exclude": ["node_modules/**"],
+	"compilerOptions": {
+		"types": ["node"]
+	}
+}
